Ensure type inference for assertThrows works properly
Issue #1394 improved the assertThrows() failure message if a lambda
expression or method reference returns a result instead of throwing an
exception. To achieve this, it introduced new variants of
assertThrows() that accept a ThrowingSupplier instead of an Executable.
However, that unfortunately causes the Java compiler to be unable to
disambiguate between the ThrowingSupplier and Executable variants when
performing type inference in certain scenarios. The result is that
certain code which previously compiled no longer compiles.

This commit fixes this issue by having ThrowingSupplier extend
Executable with a default implementation of Executable's execute()
method that delegates to ThrowingSupplier's get() method and simply
ignores the return value. When affected code is recompiled against the
new APIs, the inferred type will now be a ThrowingSupplier instead of
an Executable; however, the runtime semantics of the assertions remain
unchanged. In addition, previously compiled code should be unaffected
by these changes since the Executable variants remain intact.

Fixes: #1414