Introduce NamedArguments to name a set of arguments in a @ParameterizedTest
This is a proof of concept which demonstrates one way to achieve this.

Running NamedArgumentsSetDemo, results in the following output:

Important Files :: [1] file1=VIP, file2=path2
[2] file1=path3, file2=path4

If we decide to implement this feature, we will need to revisit the
current implementation of ParameterizedTestNameFormatter and consider
introducing a new placeholder for display names for parameterized tests
in order to allow users to specify inclusion of the "name of a set of
arguments".

See #3818