Retain original TestEngine order after pruning
Prior to this commit, LauncherDiscoveryResult.retainEngines() stored the
retained LinkedHashMap content in a HashMap, which sometimes resulted in
different iteration order if test engines were pruned.

To address that, this commit replaces the use of the Stream API and
Collectors.toMap() (which creates a HashMap) -- with the use of a
LinkedHashMap combined with entrySet().removeIf() which allows us to
ensure that the retained test engines are stored in a LinkedHashMap
that maintains the original iteration order.

Fixes #4862
Closes #4864

(cherry picked from commit a400fc6c3ff1899c2bc0e2bf350b338b55364d02)