Relocate MediaType to jupiter.api package and deprecate existing class
org.junit.jupiter.api.extension.MediaType currently causes cycles
between the `api` and `api.extension` packages in junit-jupiter-api.

Since the junit-jupiter-api artifact constitutes the public API for
JUnit Jupiter, we have traditionally not allowed package cycles within
that artifact.

Thus, in order to reduce package cycles and improve the quality of the
code base, this commit introduces a new MediaType in the
org.junit.jupiter.api package. Note that the existing MediaType now
extends the new MediaType.

To encourage users to migrate to the new supported MediaType this
commit also introduces the following deprecations.

* `org.junit.jupiter.api.extension.MediaType` is now deprecated in
  favor of the new `org.junit.jupiter.api.MediaType`.

* The `publishFile(...)` methods in `TestReporter` which accept an
  `org.junit.jupiter.api.extension.MediaType` are now deprecated in
  favor of new variants which accept an
  `org.junit.jupiter.api.MediaType`.

* The `publishFile(...)` method in `ExtensionContext` which accepts an
  `org.junit.jupiter.api.extension.MediaType` is now deprecated in
  favor of a new variant which accepts an
  `org.junit.jupiter.api.MediaType`.

Furthermore, we plan to remove the deprecated MediaType in JUnit 6.2.

See #4876
Closes #4886